remove references to watched mutation
