+++ MIGRATING.md
@@ -8,5 +8,5 @@ Please open an issue or a pull request if you feel this doc is incomplete.
 - Update React Apollo and Apollo Client to access GraphQL hooks: `npm i --save-exact apollo-client@2.6.3; npm i --save react-apollo@3.0.0`
 - `editMutation`, `newMutation` etc. are deprecated, use the new `updateFoo`, `createFoo` syntax. An error message is thrown where deprecated mutations are used to help debugging
 - No need to call `registerComponent` anymore to use Vulcan HOC. You can call them directly even if the underlying fragment is not yet registered.
-
+- Watched Mutations has been removed because it didn't work anymore, in favour to better Apollo's `update` option for mutations.
