+++ packages/vulcan-core/lib/modules/containers/cacheUpdate.js
@@ -3,6 +3,19 @@
  */
 
 
+/**
+ * Safe getter
+ * Must returns null if the document is absent (eg in case of validation failure)
+ * TODO: use this getter
+ * @param {*} mutation 
+ * @param {*} mutationName 
+ */
+const getDocumentFromMutation = (mutation, mutationName) => {
+    const mutationData = (mutation.result.data[mutationName] || {});
+    const document = mutationData.data;
+    return document;
+};
+
 // When using multi queries, we can't track all parameters, which are sadly needed
 // by cache.readQuery for optimistic updates.
 // This function can get a list of queries based on their name and should solve this issue
@@ -55,16 +68,73 @@ export const addToData = ({ queryResult, multiResolverName, document }) => {
 };
 
 
+export const removeFromData = ({ queryResult, multiResolverName, document }) => {
+    const queryData = queryResult[multiResolverName];
+    return {
+        ...queryResult,
+        [multiResolverName]: {
+            ...queryData,
+            results: queryData.results.filter(item => item._id !== document._id),
+            totalCount: Math.max(0, queryData.totalCount - 1)
+        }
+    };
+};
+
 /*
 
-Remove a document from a set
+Test if a document is matched by a given selector
 
 */
-export const removeFromSet = (queryData, document) => {
-    const newData = {
-        ...queryData,
-        results: queryData.results.filter(item => item._id !== document._id),
-        totalCount: queryData.totalCount - 1,
-    };
-    return newData;
+//export const belongsToSet = (document, selector) => {
+//  const mingoQuery = new Mingo.Query(selector);
+//  return mingoQuery.test(document);
+//};
+
+/*
+
+Test if a document is already in a result set
+
+*/
+// export const isInSet = (data, document) => data.results.find(item => item._id === document._id);
+
+/*
+
+Add a document to a set of results
+
+*/
+// export const addToSet = (queryData, document) => {
+//   const newData = {
+//     ...queryData,
+//     results: [...queryData.results, document],
+//     totalCount: queryData.totalCount + 1,
+//   };
+//   return newData;
+// };
+
+/*
+
+Update a document in a set of results
+
+*/
+// TODO: legacy, not used anymore because Apollo handles it out of the box
+/* export const updateInSet = (queryData, document) => {
+  const oldDocument = queryData.results.find(item => item._id === document._id);
+  const newDocument = { ...oldDocument, ...document };
+  const index = queryData.results.findIndex(item => item._id === document._id);
+  const newData = { ...queryData, results: [...queryData.results] }; // clone
+  newData.results[index] = newDocument;
+  return newData;
 };
+*/
+
+/*
+
+Reorder results according to a sort
+
+*/
+//export const reorderSet = (queryData, sort, selector) => {
+//  const mingoQuery = new Mingo.Query(selector);
+//  const cursor = mingoQuery.find(queryData.results);
+//  queryData.results = cursor.sort(sort).all();
+//  return queryData;
+//};
