+++ packages/vulcan-lib/lib/client/apollo-client/index.js
@@ -1,3 +1,2 @@
-export * from './updates';
 export * from './apolloClient';
 export * from './links/registerLinks';
