--- packages/vulcan-lib/lib/client/apollo-client/links/watchedMutation.js
@@ -1,12 +1,0 @@
-/**
- * Watched mutations allow to update the cache based on a MutationSuccess
- * E.g. if the user juste created a Todo, we can update the cached Todo list
- * @see https://github.com/haytko/apollo-link-watched-mutation
- */
-import { getSetting } from 'meteor/vulcan:lib';
-import WatchedMutationLink from 'apollo-link-watched-mutation';
-import { WatchedMutations } from '../updates';
-import cache from '../cache';
-const watchedMutationLink = new WatchedMutationLink(cache, WatchedMutations, getSetting('debug', undefined));
-
-export default watchedMutationLink;
